Correctly search for the best update
This commit fixes a minor bug in SearchUpdate which led to return the last improving update move rather than the best improving update move.

This commit also makes the search slightly more flexible by allowing the caller to specify the utilization threshold to validate move. This can be useful, for example, if one wants to provide allow moves that reduce the load of the selected edge and but slightly increase the max utilization.

The `SRTE.State` field is also set to private to prevents clients from mutating it.