Select demands using a roulette wheel selection
This commit changes the LGS algorithm to use roulette wheel selection for demands rather than systematically selecting the demand that contributes the most to the edge's load. This change makes the algorithm more robust and helps escape some local optima.

The roulette wheel selection is done using the new `wheels.DemandWheel` data structure which is similar to the data structure used to select edges (which is renamed `wheels.StaticWheel` as part of this commit).

Each edge maintains its own wheel of demands. Demands can be inserted/removed/updated in O(log N) where N is the number of demands in the wheel of a given edge.

This commit also makes minor code and comment improvement in different places.