Use param Beta to select demands
This commits implements demand selection using the same roulette wheel algorithm as described in the original LGS paper _Expect the Unexpected: Sub-second optimization for Segment Routing_.

To do so, the commit extends `DemandWheel` to be a roulette wheel and a map at the same time. The map part maintains a mapping from demand ID to the load/traffic the demand is sending (int64). The roulette part relies on arbitrary weights (float64) associated to the demand ID. In this case, the weight of a demand is its contributions to the utilization of an edge.

This commit also increases the initial capacity of the demand wheels from 16 to 64.